Update Submit-CertificateRequest.ps1
Changed Submit-CertificateRequest to accept either a Path or CsrContents, with CsrContents being the contents of a file at a given path. This is useful when the code is being ran in a place where there is not access file path or you have the csr contents and do not want to create a file from them just to read them back from it. I ran into this requirement running PowerShell on a MID server as part of a workflow and creating a file for the $Path variable was much more difficult than just passing in the contents.